Se agrega sweetAlert, paginado, y guard para el login

+++ Tesla-Front/tesla-client-new/src/app/component/listarCategoria/listarCategoria.component.ts
@@ -5,6 +5,8 @@ import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { HttpErrorResponse } from '@angular/common/http';
+import { SweetAlert2Service } from '../../servicio/sweetAlert2.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -20,8 +22,15 @@ export class ListarCategoriaComponent implements OnInit {
   showModal: boolean = false;
   editForm: FormGroup;  // Formulario reactivo para edición
 
+  //Paginado
+  categoriasPorPagina: any[] = [];  // Productos que se mostrarán en la página actual
+  paginaActual: number = 1;
+  registrosPorPagina: number = 7;
+
   constructor(private categoriaService: CategoriaService,
-    private formBuilder: FormBuilder) {
+    private formBuilder: FormBuilder,
+    private sweetAlertService: SweetAlert2Service
+  ) {
     // Inicializamos el formulario reactivo con validaciones
     this.editForm = this.formBuilder.group({
       id: [null],  // Agregar el campo id
@@ -39,6 +48,7 @@ export class ListarCategoriaComponent implements OnInit {
     this.categoriaService.getCategoria().subscribe({
       next: (response) => {
         this.categorias = response;  // Asignar las categorías obtenidas
+        this.actualizarPagina();
 
       },
       error: (error) => {
@@ -67,30 +77,74 @@ export class ListarCategoriaComponent implements OnInit {
   // Método para guardar los cambios de la categoría editada
   guardarCategoria(): void {
     if (this.editForm.valid) {
-      const categoriaEditada = this.editForm.value;  // Obtener valores del formulario reactivo
+      if (this.editForm.valid) {
+        const categoriaEditada = this.editForm.value;  // Obtener valores del formulario reactivo
 
-      this.categoriaService.putCategoria(categoriaEditada).subscribe({
+        // Mostrar el mensaje de confirmación antes de guardar los cambios
+        this.sweetAlertService.showConfirmMessage('¿Estás seguro?', '¿Quieres guardar los cambios realizados en esta categoría?')
+          .then((result) => {
+            if (result.isConfirmed) {
+              // El usuario confirmó, proceder a guardar la categoría
+              this.categoriaService.putCategoria(categoriaEditada).subscribe({
+                next: () => {
+                  this.getCategoria();  // Actualizar lista de categorías
+                  this.cerrarModal();   // Cerrar modal
+                },
+                error: (error: HttpErrorResponse) => {
+                  console.error('Error al guardar la categoría', error);
+                }
+              });
+            } else {
+              // El usuario canceló, no hacer nada
+              console.log('Se canceló la acción');
+            }
+          });
+      } else {
+        console.log('Formulario no válido');
+      }
+  }
+}
+
+eliminarCategoria(id: number): void {
+  this.sweetAlertService.showConfirmMessage('¿Estás seguro?', '¿Deseas eliminar esta categoría?').then((result) => {
+    if (result.isConfirmed) {
+      // Llamar al servicio para eliminar la categoría si se confirma
+      this.categoriaService.deleteCategoria(id).subscribe({
         next: () => {
-          this.getCategoria();  // Actualizar lista de categorías
-          this.cerrarModal();   // Cerrar modal
+          this.getCategoria();  // Recargar lista de categorías después de eliminar
+          Swal.fire('Eliminado', 'La categoría ha sido eliminada', 'success');  // Mostrar mensaje de éxito
         },
         error: (error: HttpErrorResponse) => {
-          console.error('Error al guardar la categoría', error);
+          console.error('Error al eliminar la categoría', error);
+          Swal.fire('Error', 'No se pudo eliminar la categoría', 'error');  // Mensaje de error en caso de fallo
         }
       });
     }
+  });
+}
+
+
+  // Función para actualizar los productos que se deben mostrar en la página actual
+  actualizarPagina() {
+    const inicio = (this.paginaActual - 1) * this.registrosPorPagina;
+  const fin = inicio + this.registrosPorPagina;
+  this.categoriasPorPagina = this.categorias.slice(inicio, fin);  // Filtrar productos para la página actual
   }
 
-  // Método para eliminar una categoría
-  eliminarCategoria(id: number): void {
-    this.categoriaService.deleteCategoria(id).subscribe({
-      next: () => {
-        this.getCategoria();  // Recargar lista de categorías después de eliminar
-      },
-      error: (error: HttpErrorResponse) => {
-        console.error('Error al eliminar la categoría', error);
-      }
-    });
+  // Cambiar la página
+  cambiarPagina(pagina: number) {
+    this.paginaActual = pagina;
+    this.actualizarPagina();
+  }
+
+  // Verificar si hay una página siguiente
+  tienePaginaSiguiente(): boolean {
+    return this.paginaActual < Math.ceil(this.categorias.length / this.registrosPorPagina);
+  }
+
+  // Verificar si hay una página anterior
+  tienePaginaAnterior(): boolean {
+    return this.paginaActual > 1;
   }
 
 }
